refactor: delete batch_anonymizer
