+++ great_expectations/validator/validator.py
@@ -1710,9 +1710,7 @@ set as active.
                     # noinspection PyProtectedMember
                     handler.send_usage_message(
                         event="data_asset.validate",
-                        event_payload=handler._batch_anonymizer.anonymize_batch_info(
-                            self
-                        ),
+                        event_payload=handler._anonymizer.anonymize_batch_info(self),
                         success=False,
                     )
                 return ExpectationValidationResult(success=False)
@@ -1822,7 +1820,7 @@ set as active.
                 # noinspection PyProtectedMember
                 handler.send_usage_message(
                     event="data_asset.validate",
-                    event_payload=handler._batch_anonymizer.anonymize_batch_info(self),
+                    event_payload=handler._anonymizer.anonymize_batch_info(self),
                     success=False,
                 )
             raise
@@ -1835,7 +1833,7 @@ set as active.
             # noinspection PyProtectedMember
             handler.send_usage_message(
                 event="data_asset.validate",
-                event_payload=handler._batch_anonymizer.anonymize_batch_info(self),
+                event_payload=handler._anonymizer.anonymize_batch_info(self),
                 success=True,
             )
         return result
